+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Movies/Create.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RazorPagesMovie.Pages.Movie.CreateModel
+@model RazorPagesMovie.Pages.Movies.CreateModel
 
 @{
     ViewData["Title"] = "Create";
