+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Movies/Delete.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RazorPagesMovie.Pages_Movie.DeleteModel
+@model RazorPagesMovie.Pages.Movies.DeleteModel
 
 @{
     ViewData["Title"] = "Delete";
