+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Movies/Delete.cshtml.cs
@@ -7,17 +7,18 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using RazorPagesMovie.Models;
 
-namespace RazorPagesMovie.Pages_Movie
+namespace RazorPagesMovie.Pages.Movies
 {
-    public class DetailsModel : PageModel
+    public class DeleteModel : PageModel
     {
         private readonly RazorPagesMovie.Models.MovieContext _context;
 
-        public DetailsModel(RazorPagesMovie.Models.MovieContext context)
+        public DeleteModel(RazorPagesMovie.Models.MovieContext context)
         {
             _context = context;
         }
 
+        [BindProperty]
         public Movie Movie { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
@@ -34,6 +35,24 @@ namespace RazorPagesMovie.Pages_Movie
                 return NotFound();
             }
             return Page();
+        }
+
+        public async Task<IActionResult> OnPostAsync(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            Movie = await _context.Movie.FindAsync(id);
+
+            if (Movie != null)
+            {
+                _context.Movie.Remove(Movie);
+                await _context.SaveChangesAsync();
+            }
+
+            return RedirectToPage("./Index");
         }
     }
 }
