+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Movies/Details.cshtml
@@ -1,13 +1,12 @@
 @page
-@model RazorPagesMovie.Pages_Movie.DeleteModel
+@model RazorPagesMovie.Pages.Movies.DetailsModel
 
 @{
-    ViewData["Title"] = "Delete";
+    ViewData["Title"] = "Details";
 }
 
-<h2>Delete</h2>
+<h2>Details</h2>
 
-<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Movie</h4>
     <hr />
@@ -37,10 +36,8 @@
             @Html.DisplayFor(model => model.Movie.Price)
         </dd>
     </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Movie.ID" />
-        <input type="submit" value="Delete" class="btn btn-default" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
+</div>
+<div>
+    <a asp-page="./Edit" asp-route-id="@Model.Movie.ID">Edit</a> |
+    <a asp-page="./Index">Back to List</a>
 </div>
