+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Movies/Edit.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RazorPagesMovie.Pages_Movie.EditModel
+@model RazorPagesMovie.Pages.Movies.EditModel
 
 @{
     ViewData["Title"] = "Edit";
