+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/_Layout.cshtml
@@ -26,7 +26,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                <a asp-page="/Movie/Index" class="navbar-brand">RpMovie</a>
+                <a asp-page="/Movies/Index" class="navbar-brand">RpMovie</a>
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
