+++ aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Pages/Movie/Create.cshtml.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using RazorPagesMovie.Models;
 using System.Threading.Tasks;
 
-namespace RazorPagesMovie.Pages_Movie
+namespace RazorPagesMovie.Pages.Movies
 {
     public class CreateModel : PageModel
     {
@@ -19,13 +19,20 @@ namespace RazorPagesMovie.Pages_Movie
 
         public IActionResult OnGet()
         {
+            Movie = new Movie
+            {
+                Title = "The Good, the bad, and the ugly",
+                Genre = "Western",
+                Price = 1.19M,
+                ReleaseDate = DateTime.Now
+            };
             return Page();
         }
 
         [BindProperty]
         public Movie Movie { get; set; }
 
-#region snippetPost
+        #region snippetPost
         public async Task<IActionResult> OnPostAsync()
         {
             if (!ModelState.IsValid)
@@ -38,7 +45,7 @@ namespace RazorPagesMovie.Pages_Movie
 
             return RedirectToPage("./Index");
         }
-#endregion
+        #endregion
     }
 }
 #endregion
