--- aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Pages/Movie/Delete.cshtml.cs
@@ -1,58 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using RazorPagesMovie.Models;
-
-namespace RazorPagesMovie.Pages_Movie
-{
-    public class DeleteModel : PageModel
-    {
-        private readonly RazorPagesMovie.Models.MovieContext _context;
-
-        public DeleteModel(RazorPagesMovie.Models.MovieContext context)
-        {
-            _context = context;
-        }
-
-        [BindProperty]
-        public Movie Movie { get; set; }
-
-        public async Task<IActionResult> OnGetAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Movie = await _context.Movie.SingleOrDefaultAsync(m => m.ID == id);
-
-            if (Movie == null)
-            {
-                return NotFound();
-            }
-            return Page();
-        }
-
-        public async Task<IActionResult> OnPostAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            Movie = await _context.Movie.FindAsync(id);
-
-            if (Movie != null)
-            {
-                _context.Movie.Remove(Movie);
-                await _context.SaveChangesAsync();
-            }
-
-            return RedirectToPage("./Index");
-        }
-    }
-}
