--- aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Pages/Movie/Details.cshtml
@@ -1,44 +1,0 @@
-@page
-@model RazorPagesMovie.Pages_Movie.DetailsModel
-
-@{
-    ViewData["Title"] = "Details";
-}
-
-<!-- a comment-->
-<h2>Details</h2>
-
-<div>
-    <h4>Movie</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Movie.Title)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Movie.Title)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Movie.ReleaseDate)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Movie.ReleaseDate)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Movie.Genre)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Movie.Genre)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Movie.Price)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Movie.Price)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-page="./Edit" asp-route-id="@Model.Movie.ID">Edit</a> |
-    <a asp-page="./Index">Back to List</a>
-</div>
