+++ aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Pages/Movie/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model RazorPagesMovie.Pages_Movie.IndexModel
+@model RazorPagesMovie.Pages.Movies.IndexModel
 
 @{
     ViewData["Title"] = "Index";
